dms_api/requirements.txt
# DMS API Gateway - FastAPI Requirements

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0

# File Upload
python-multipart>=0.0.6

# OCR Engine
cnocr>=2.2.0
onnxruntime>=1.16.0
Pillow>=10.0.0
numpy>=1.24.0

# Development & Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Optional: YAML configuration support
pyyaml>=6.0
